refactor: move URI method for RDWS to separate uri.go file
